avoid \@ signs (representing nulls) in log files
The byte consisting of one null character differs from the 1-character string
consisting of only the terminating byte; the latter is a pointer to a non-null
address. Running gevolution and storing the stdout+stderr log file with the old version
of this file gives null characters in the output, which looks bad when reading
the log, and is probably not a safe programming practice (hidden characters
in a plain text file).